Run the IsaacSim interactive worker as a plain subprocess
Kit segfaults in _start_app when booted inside a multiprocessing-spawned
child (first from the spawn re-import pulling in viser's embreex stack,
which collides with Kit's bundled embree; a clean child still crashed).
Replace the multiprocessing worker with a regular subprocess running
_isaacsim_interactive_worker.py, speaking the same message protocol over a
localhost socket Connection. InteractiveDemo gains a worker_factory hook
returning (proc, conn); the gym backend keeps the original Pipe/Process
path. Verified: worker reaches ready (claw_hammer/swing_down) in ~31s.
